backend/requirements.txt
django>=4.2,<5
djangorestframework>=3.14,<4
django-cors-headers>=4.3,<5
pytest>=8.0,<9
pytest-django>=4.8,<5
black>=24.0,<25
ruff>=0.5,<1
